IDT: some fixes to ensure basic working condition
I followed Jame Molloy's tutorial on IDT, but was having a hard time to
resolve a bug which happens after an 'INT $3' instuction. I don't
understand the nature of the bug, which crashes my kernel by accessing
somewhere outside the RAM or ROM. Debugger wasn't helping, so I'm left
on my own.

I decided to consult the Intel's i386 system programming manual, and
realized that the way that the tutorial passes arguments onto the stack
is a bit different from what's written in the manual. I'm not saying the
tutorial is definitely wrong, because I could not find any errors by
inspecting its code, namely how it pushes all registers to stack then
pops them off. But what I did is, I commented out the part of code that
shouldn't be there per i386 manual, then all of a sudden the ISR worked
(well, not in any useful fasion, but at lease it does not crash
anymore).

So I'm keeping this change for now. I understand that the tutorial wants to
utilize all those registers' info later in the ISR, but I decided not to
pass them for now, simply for not crashing the kernel. It seems like a
bit of progression for me, from copying the tutorial verbatim, to
actually check the manual and resolve problems on my own. Kudos to
myself.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>