overhaul: flatten kernel/ directory
Overhaul is a series of changes that aim to bring the whole project into
a flattened directory. The motivation is to cut away unnecessary mental
burden created by nested directories. This is not to say that
directories are useless, but they should only be added when it's
absolutely beneficial/necessary. Before that, we can keep things simple
in a single, flattened directory. This is a modification based on the
Meaty Skeleton template provided at osdev.

This commit is the first step, which flattened the kernel/ directory.
The next steps involve flattening the libc/ directory, and eventually
bringing everything into the same root directory.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>